requirements.txt
Complete output from command git config --get-regexp remote\..*\.url:

----------------------------------------
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.0
numpy==1.15.2
Pillow==5.3.0
pyparsing==2.2.2
python-dateutil==2.7.3
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
## !! Could not determine repository location
svae==0.0.0
telepyth==0.1.6
torch==0.4.1.post2
torchvision==0.2.1
tqdm==4.27.0
